Fix file source seek prefill int overflow and frame alignment
prefillFileWaterfall computed lineSeconds as (int)lineIndex * fftInterval,
which overflows int at ~214s on a 10 MHz file, reading the wrong position
and garbling the waterfall. Use double arithmetic there, and round the
seek to the nearest whole frame (llround) so it always lands on an I/Q
frame boundary. Also serialize computeFFT, which shares one FFTW plan
between the live FFT and the seek prefill.

+++ core/src/gui/main_window.cpp
@@ -288,7 +288,7 @@ void MainWindow::prefillFileWaterfall(double seekTime) {
         // Fill lines oldest -> newest so setHistory() can put the newest on top.
         for (int k = 0; k < count; k++) {
             int lineIndex = newestLine - (count - 1 - k);
-            double lineSeconds = (double)(lineIndex * fftInterval) / sr;
+            double lineSeconds = ((double)lineIndex * (double)fftInterval) / sr;
             int got = sigpath::sourceManager.readSamples(lineSeconds, samples, nz);
             if (got < nz) {
                 memset(samples + got, 0, (nz - got) * sizeof(dsp::complex_t));
